rename sequelize file to dbHelper, remove redundant comment, synchronise schema without force parameter
